jwt_manager/requirements.txt
cryptography >= 3.4.8
PyJWT >= 2.3.0
